@@ -7,8 +7,8 @@ import com.web.stard.domain.member.domain.Member;
 import com.web.stard.global.domain.BaseEntity;
 import lombok.*;
 
-import javax.persistence.*;
-import javax.validation.constraints.NotNull;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
 
 @Entity
 @Table
