@@ -6,8 +6,8 @@ import com.web.stard.domain.board.study.domain.StudyPost;
 import com.web.stard.domain.member.domain.Member;
 import lombok.*;
 
-import javax.persistence.*;
-import javax.validation.constraints.NotNull;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
 
 @Entity
 @Table
