@@ -1,11 +1,10 @@
 package com.web.stard.domain.board.study.domain;
 
-import com.sun.istack.NotNull;
+import jakarta.validation.constraints.NotNull;
 import com.web.stard.domain.member.domain.Member;
-import com.web.stard.domain.board.study.domain.Study;
 import lombok.*;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 @Table
