@@ -1,11 +1,10 @@
 package com.web.stard.domain.board.study.domain;
 
-import com.sun.istack.NotNull;
-import com.web.stard.domain.board.study.domain.Study;
+import jakarta.validation.constraints.NotNull;
 import lombok.*;
 import org.springframework.format.annotation.DateTimeFormat;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.time.LocalDateTime;
 
 @Entity
