@@ -3,8 +3,8 @@ package com.web.stard.domain.board.study.domain;
 import com.web.stard.domain.member.domain.Member;
 import lombok.*;
 
-import javax.persistence.*;
-import javax.validation.constraints.NotNull;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
 
 @Entity
 @Builder
