@@ -2,16 +2,13 @@ package com.web.stard.domain.member.api;
 
 import com.web.stard.domain.member.dto.EmailDto;
 import com.web.stard.domain.member.application.EmailService;
-import com.web.stard.domain.member.dto.PasswordUpdateDto;
 import com.web.stard.domain.member.dto.ResetPasswordResponse;
 import com.web.stard.domain.member.application.MemberService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
-import javax.mail.MessagingException;
+import jakarta.mail.MessagingException;
 
 @RestController
 @RequiredArgsConstructor
